requirements.txt
django
djangorestframework
django-rest-swagger
djangorestframework-gis